Unit test updates

+++ src/main/java/com/bbthechange/inviter/service/JwtService.java
@@ -64,8 +64,8 @@ public class JwtService {
         }
     }
     
-    public long getAccessTokenExpirationSeconds() {
-        return ACCESS_TOKEN_EXPIRATION / 1000;
+    public int getAccessTokenExpirationSeconds() {
+        return (int) (ACCESS_TOKEN_EXPIRATION / 1000);
     }
 
     private Claims extractClaims(String token) {
